chore: relocate settings service
